+++ product-api/sdk/models/product.go
@@ -14,7 +14,7 @@ import (
 	"github.com/go-openapi/validate"
 )
 
-// Product Product Product Product defines the structure for an API product
+// Product Product Product defines the structure for an API product
 //
 // swagger:model Product
 type Product struct {
@@ -35,7 +35,7 @@ type Product struct {
 	// the price for the product
 	// Required: true
 	// Minimum: 0.01
-	Price *float32 `json:"price"`
+	Price *float64 `json:"price"`
 
 	// the SKU for the product
 	// Required: true
@@ -116,7 +116,7 @@ func (m *Product) validatePrice(formats strfmt.Registry) error {
 		return err
 	}
 
-	if err := validate.Minimum("price", "body", float64(*m.Price), 0.01, false); err != nil {
+	if err := validate.Minimum("price", "body", *m.Price, 0.01, false); err != nil {
 		return err
 	}
